+++ examples/ttf.py
@@ -81,7 +81,7 @@ def main() -> None:
             for event in tcod.event.wait():
                 if isinstance(event, tcod.event.Quit):
                     raise SystemExit()
-                if isinstance(event, tcod.event.WindowResized) and event.type == "WINDOWSIZECHANGED":
+                if isinstance(event, tcod.event.WindowResized) and event.type == "WindowSizeChanged":
                     # Resize the Tileset to match the new screen size.
                     context.change_tileset(
                         load_ttf(